+++ supercharged_dart_plus/lib/list/list_object.dart
@@ -1,4 +1,4 @@
-part of supercharged_dart_plus;
+part of '../supercharged_dart_plus.dart';
 
 /// Supercharged extensions on [List].
 extension ListSC<T> on List<T> {