+++ supercharged_dart_plus/lib/map/map.dart
@@ -1,4 +1,4 @@
-part of supercharged_dart_plus;
+part of '../supercharged_dart_plus.dart';
 
 /// Supercharged extensions on [Map].
 extension MapSC<K, V> on Map<K, V> {