+++ supercharged_plus/example/lib/main.dart
@@ -16,41 +16,46 @@ class MyApp extends StatelessWidget {
           title: const Text('Supercharged Plus Example'),
           backgroundColor: '#3F51B5'.toColor(),
         ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Container(
-                width: 150,
-                height: 100,
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.circular(12),
-                  boxShadow: const [
-                    BoxShadow(
-                      color: Colors.black26,
-                      blurRadius: 10,
-                      offset: DeltaOffset.rightBottom(6.0, 4.0),
-                    ),
-                  ],
-                ),
-                child: Center(
+        body: Builder(
+          builder: (context) {
+            return Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Container(
+                  width: 150,
+                  height: 100,
+                  decoration: BoxDecoration(
+                    color: Colors.white,
+                    borderRadius: BorderRadius.circular(12),
+                    boxShadow: const [
+                      BoxShadow(
+                        color: Colors.black26,
+                        blurRadius: 10,
+                        offset: DeltaOffset.rightBottom(6.0, 4.0),
+                      ),
+                    ],
+                  ),
                   child: Text(
                     'DeltaOffset',
                     style: TextStyle(
                       color: '#212121'.toColor(),
                       fontWeight: FontWeight.bold,
                     ),
-                  ),
+                  ).center(),
                 ),
-              ),
-              const SizedBox(height: 24),
-              Text(
-                'Sorted: ${[3, 1, 4, 2]..sortAsc()}',
-                style: const TextStyle(fontSize: 16),
-              ),
-            ],
-          ),
+                const SizedBox(height: 24),
+                Text(
+                  'Sorted: ${[3, 1, 4, 2]..sortAsc()}',
+                  style: const TextStyle(fontSize: 16),
+                ),
+                const SizedBox(height: 12),
+                Text(
+                  'Is Dark Mode: ${context.isDarkMode}',
+                  style: const TextStyle(fontSize: 16),
+                ),
+              ],
+            ).center().paddingAll(16);
+          },
         ),
       ),
     );